Move shared fixtures to conftest
